--- src/main/java/SK_3team/example/planner/util/JwtUtil.java
@@ -1,55 +1,0 @@
-package SK_3team.example.planner.util;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jws;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import javax.crypto.SecretKey;
-import java.nio.charset.StandardCharsets;
-import java.util.Date;
-
-@Component
-public class JwtUtil {
-    // application.properties에서 JWT Secret Key를 로드
-    @Value("${jwt.secret.key}")
-    private String secretKey;
-
-    private SecretKey getSigningKey() {
-        return Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8));
-    }
-
-    public Claims extractAllClaims(String token) {
-        return Jwts.parserBuilder()
-                .setSigningKey(getSigningKey())
-                .build()
-                .parseClaimsJws(token)
-                .getBody();
-    }
-
-
-    // 토큰에서 사용자 ID 추출 (예시로 "userId" 클레임을 가정)
-    public Long getUserIdFromToken(String token) {
-        Claims claims = extractAllClaims(token);
-        // 클레임에서 "userId" 키로 저장된 값을 Long 타입으로 반환합니다.
-        // 실제 JWT 생성 시 어떤 이름으로 사용자 ID를 저장했는지 확인해야 합니다.
-        return Long.parseLong(claims.get("userId", String.class));
-    }
-    public boolean isTokenExpired(String token) {
-        return extractAllClaims(token).getExpiration().before(new Date());
-    }
-    public boolean validateToken(String token) {
-        try {
-            Jws<Claims> claimsJws = Jwts.parserBuilder()
-                    .setSigningKey(getSigningKey())
-                    .build()
-                    .parseClaimsJws(token);
-
-            return !claimsJws.getBody().getExpiration().before(new Date());
-        } catch (Exception e) {
-            return false;
-        }
-    }
-}
